Added the all new ovamorph model and texture, as well as new ovamorph actions. Refactored several things to match AIRI's changes to CoordData.

+++ src/main/java/com/arisux/avp/entities/mob/EntityDrone.java
@@ -148,8 +148,8 @@ public class EntityDrone extends EntityXenomorph
 
 							if (Entities.canCoordBeSeenBy(this, coord) && block.isOpaqueCube())
 							{
-								this.getNavigator().setPath(this.worldObj.getEntityPathToXYZ(this, coord.posX, coord.posY, coord.posZ, 128, true, true, true, true), 0.8D);
-								this.worldObj.setBlock(coord.posX, coord.posY, coord.posZ, AliensVsPredator.blocks().terrainHiveResin);
+								this.getNavigator().setPath(this.worldObj.getEntityPathToXYZ(this, (int) coord.posX, (int) coord.posY, (int) coord.posZ, 128, true, true, true, true), 0.8D);
+								this.worldObj.setBlock((int) coord.posX, (int) coord.posY, (int) coord.posZ, AliensVsPredator.blocks().terrainHiveResin);
 
 								TileEntity tileEntity = coord.getTileEntity(this.worldObj);
